Refactor out sections and apply section and title styles
