+++ frontends/mit-open/src/pages/HomePage/UpcomingCoursesSection.tsx
@@ -1,0 +1,47 @@
+import React from "react"
+import { Container, styled, Typography } from "ol-components"
+import TabbedCarousel, {
+  TabbedCarouselProps,
+} from "@/page-components/TabbedCarousel/TabbedCarousel"
+
+const Section = styled.section`
+  padding: 80px 0px;
+  overflow: auto;
+`
+
+const UPCOMING_COURSES_CAROUSEL: TabbedCarouselProps["config"] = [
+  {
+    label: "All",
+    pageSize: 4,
+    data: {
+      type: "resources",
+      params: { resource_type: ["course"], limit: 12, sortby: "upcoming" },
+    },
+  },
+  {
+    label: "Professional",
+    pageSize: 4,
+    data: {
+      type: "resources",
+      params: {
+        professional: true,
+        resource_type: ["course"],
+        limit: 12,
+        sortby: "upcoming",
+      },
+    },
+  },
+]
+
+const UpcomingCoursesSection: React.FC = () => {
+  return (
+    <Section>
+      <Container>
+        <Typography variant="h2">Upcoming Courses</Typography>
+        <TabbedCarousel config={UPCOMING_COURSES_CAROUSEL} />
+      </Container>
+    </Section>
+  )
+}
+
+export default UpcomingCoursesSection
